% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{convert_bbox}
\alias{convert_bbox}
\title{Convert between BBox Formats}
\usage{
convert_bbox(
  bbox,
  from,
  to,
  image = NULL,
  img_width = NULL,
  img_height = NULL,
  value_type = "absolute"
)
}
\arguments{
\item{bbox}{character vector. Bounding box values.}

\item{from, to:}{\itemize{
\item \code{from} character vector. Bounding box format to convert from.
\item \code{to} character vector. Bounding box format to convert to.
\item Options: "xyxy", "xywh", and "centerwh"
}}

\item{image}{character or magick-image. The image to extract width and height from.
Only used if converting between value types (\code{value_type} is opposite of the provided
bounding box). Value type of the input bounding box is inferred. Mutually exclusive
with \code{img_width} and \code{img_height}.}

\item{img_width, img_height}{numeric. The width and height of the image this bounding box
belongs to. Only used if converting between value types (\code{value_type} is opposite of the
provided bounding box). Value type of the input bounding box is inferred. Mutually exclusive
mutually exclusive with \code{image}.}

\item{value_type}{character. Type of values to use for the converted bounding box measures.
Either "absolute" for pixel values or "relative" for scaled values.}
}
\value{
a numberic vector.
}
\description{
Converts bounding box data between three common formats: xyxy, xywh, and centerwh.
WARNING: This function is experimental, math behind some conversions may be wrong
and has not yet been validated.
}
