% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ena24detection_subset}
\alias{ena24detection_subset}
\title{Random Subset of ena24detection files}
\format{
An object of class \code{character} of length 500.
}
\usage{
ena24detection_subset
}
\description{
A random Subset of 500 image files from the ena24detection dataset to use for testing.
}
\keyword{datasets}
