% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{img_apply}
\alias{img_apply}
\title{Apply a Summary function to an Image}
\usage{
img_apply(image, fun, bbox = NULL, ...)
}
\arguments{
\item{image}{character or magick-image. Image to extract information from.}

\item{bbox}{numeric vector. Bounding box values. Formatted as xmin, ymin, xmax, ymax.
Use \code{\link[=convert_bbox]{convert_bbox()}} to convert from other formats.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
varies based on function applied.
}
\description{
Applies a function to the extracted numeric matrix of values behind an image
and returns the result. Functions are applied using \code{apply()} across channels.
#'
}
