% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{img_diff}
\alias{img_diff}
\title{Calculate Differences Between Two Images}
\usage{
img_diff(image1, image2, file = NULL, bbox = NULL, threshold = NULL)
}
\arguments{
\item{image1, image2}{character or magick-image. Images to be compared.}

\item{file}{character. File path to the output image.}

\item{bbox}{numeric vector. Bounding box values. Formatted as xmin, ymin, xmax, ymax.
Use \code{\link[=convert_bbox]{convert_bbox()}} to convert from other formats.}

\item{threshold}{numeric. The image threshold that should be applied between 0 and 1.
All values below the threshold will be converted to 0.}
}
\value{
a magick image.
}
\description{
Calculates the absolute pixel difference between two images.
}
