% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{img_difference}
\alias{img_difference}
\title{Calculate Differences Between Two Images}
\usage{
img_difference(
  image1,
  image2,
  threshold = NULL,
  file = NULL,
  overwrite = FALSE,
  bbox = NULL
)
}
\arguments{
\item{image1, image2}{character or magick-image. Images to be compared.}

\item{threshold}{numeric. The image threshold that should be applied between 0 and 1.
All values below the threshold will be converted to 0.}

\item{file}{character. File path to the output image.}

\item{bbox}{numeric vector. Bounding box values. Formatted as xmin, ymin, xmax, ymax.
Use \code{\link[=convert_bbox]{convert_bbox()}} to convert from other formats.}
}
\value{
a magick image.
}
\description{
Calculates the absolute pixel difference between two images.
}
