% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{img_draw_bbox}
\alias{img_draw_bbox}
\title{Add Bounding Box(es) to Images}
\usage{
img_draw_bbox(
  image,
  bbox,
  file = NULL,
  overwrite = FALSE,
  validate_json = getOption("camRa.validate_json", default = TRUE),
  ...
)
}
\arguments{
\item{image}{character or magick-image. The image to extract text from.}

\item{bbox, json}{Pick one of \code{bbox} and \code{json}:
\itemize{
\item \code{bbox} character vector. Bounding box values. Formatted as xmin, ymin, xmax, ymax
using absolute pixel values (Top-Left Bottom-Right format). Use \code{NA} to get text from
the entire image.
\item \code{json}character vector or nested list object from \code{\link[jsonlite:read_json]{jsonlite::read_json()}}.
JSON data with bounding box info to pull for the image.
}}

\item{file}{character. File to write image to. Use \code{NA} to skip writing to a file.}

\item{validate_json}{boolean. If JSON data formatted as nested lists should be validated.
This can prevent unexpected errors if the parameter is a list, but not JSON but may increase runtime.}

\item{...}{additional arguments passed to \code{\link[=rect]{rect()}}.}
}
\value{
a magick image.
}
\description{
Overlays bounding box(es) over images and saves this as a new image.
}
