% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{img_extract_text}
\alias{img_extract_text}
\title{Extract Text from Image Using OCR}
\usage{
img_extract_text(image, bbox = NULL, file = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{image}{character or magick-image. The image to extract text from.}

\item{bbox}{numeric vector. Bounding box values. Formatted as xmin, ymin, xmax, ymax.
Use \code{\link[=convert_bbox]{convert_bbox()}} to convert from other formats.
using absolute pixel values (Top-Left Bottom-Right format). Positions are measured from
the top left corder of the image. Use \code{NULL} to get text from the entire image.}

\item{...}{additional arguments passed to OCR engine as \code{\link[tesseract:tesseract]{tesseract::tesseract()}}.}
}
\value{
a character.
}
\description{
Extracts text from an image using optical character recognition (OCR). OCR is
notably unreliable for some camera types, text extraction should be used as a
last resort and may not be accurate. Check the results of test images before
trusting OCR text outputs.
}
