% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{img_luminosity}
\alias{img_luminosity}
\title{Calculate Luminosity/Brightness of an Image}
\usage{
img_luminosity(image, method = "grayscale", file = NULL, bbox = NULL)
}
\arguments{
\item{image}{character or magick-image. Image to extract information from.}

\item{method}{character. The method used to calculate luminosity/brightness.
Options are "greyscale" or "grayscale" to get a greyscale version of an RGB image, "LAB" to extract
the lightness (L) channel from a LAB colorspace, "HSV" to extract the value (V)
channel from a HVS colorspace, "YCbCr" for the luminance channel (Y) from a YCbCr
color space, or "YIQ" for the luma (Y) channel in YIQ colorspace.}

\item{file}{character. File path to the output image.}

\item{bbox}{numeric vector. Bounding box values. Formatted as xmin, ymin, xmax, ymax.
Use \code{\link[=convert_bbox]{convert_bbox()}} to convert from other formats.}
}
\value{
a magick image.
}
\description{
Extracts luminance information from a specified colorspace and returns the
the appropriate channel.
}
