% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LILA.R
\name{LILA_download}
\alias{LILA_download}
\title{Download LILA Image Files}
\usage{
LILA_download(dataset, files, dir, quiet = FALSE)
}
\arguments{
\item{dataset}{character. Name of the dataset to download images from.
Can be either the short name or standard name. See \code{\link[=LILA_list_datasets]{LILA_list_datasets()}}.}

\item{files}{character or character vector. Partial path(s) or file name(s) of all images to be downloaded.
Files may have partial paths included here (such as \code{folder1/imgname.JPG}) depending on
how buckets are organized. If all files are stored in the same "folder" through S3, you
only need the file name with its extension. File names are typically available from JSON
metadata files hosted on the LILA website or via \code{\link[=LILA_list_files]{LILA_list_files()}}.}

\item{dir}{character. Directory where images will be saved.}

\item{quiet}{boolean. If information on file downloads should be printed to the console.}
}
\value{
\code{NULL}
}
\description{
Downloads subsets of image files from AWS S3 buckets associated with LILA datasets.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
#Download example subset files
LILA_download_files(
  dataset = "ena24detection",
  files = ena24detection_img_files$subset,
  dir = getwd()
)
\dontshow{\}) # examplesIf}
}
