% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LILA.R
\name{LILA_list_datasets}
\alias{LILA_list_datasets}
\title{Get a List of Datasets from LILA BC}
\usage{
LILA_list_datasets(quiet = FALSE, only_nonzip = TRUE)
}
\arguments{
\item{quiet}{logical. If warnings should be shown.}

\item{only_nonzip}{logical. If the returned datasets should be only ones
available file-by-file (not exclusively as compressed files).}
}
\value{
A data frame with two columns: \code{DatasetName} and \code{DatasetShortName}.
}
\description{
Provides a dataframe with names (long and short) used for LILA datasets.
Can be used to find valid values for other LILA functions.
}
