% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaDetector.R
\name{megadet_filter_json}
\alias{megadet_filter_json}
\title{Filter MegaDetector/SpeciesNet JSON Files by Image}
\usage{
megadet_filter_json(
  dir,
  json,
  file = NA,
  validate_json = getOption("camRa.validate_json", default = TRUE)
)
}
\arguments{
\item{dir}{character. Directory to be used to filter the JSON data. This folder should
not be within the relative path of the \code{file} JSON tag, but should directly contain any
folders that appear there.}

\item{json}{character vector or nested list object from \code{\link[jsonlite:read_json]{jsonlite::read_json()}}.
The JSON file or loaded JSON data to filter on.}

\item{file}{character. File to write JSON data to. Use \code{NA} to skip writing to a file.}

\item{validate_json}{boolean. If JSON data formatted as nested lists should be validated.
This can prevent unexpected errors if the parameter is a list, but not JSON but may increase runtime.}
}
\value{
a nested list object with filtered JSON data.
}
\description{
Filters MegaDetector/SpeciesNet JSON files to only data for images in the directory specified.
}
