% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaDetector.R
\name{megadet_flatten_json}
\alias{megadet_flatten_json}
\title{Flatten MegaDetector/SpeciesNet JSON Detection Data}
\usage{
megadet_flatten_json(
  json,
  map_names = TRUE,
  validate_json = getOption("camRa.validate_json", default = TRUE)
)
}
\arguments{
\item{json}{character vector or nested list object from \code{\link[jsonlite:read_json]{jsonlite::read_json()}}.
The JSON file or loaded JSON data to filter on.}

\item{map_names}{boolean. If MegaDetector and SpeciesNet names should be mapped from
numeric values to the values provided in the \code{detection_categories} and
\code{classification_categories} tags. These are typically gross (human, animal, vehicle) and
taxonomic classifications respectively.}

\item{validate_json}{boolean. If JSON data formatted as nested lists should be validated.
This can prevent unexpected errors if the parameter is a list, but not JSON but may increase runtime.}
}
\value{
a dataframe with all data from the \code{images} tag of the JSON file.
}
\description{
Flattens MegaDetector/SpeciesNet JSON files to a table format. Data is split by detection, with
images potentially having more than one row of data if more than one detection was found. Images
with no detections are not included here.
}
