% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaDetector.R
\name{megadet_get_counts}
\alias{megadet_get_counts}
\title{Count Number of Instances of MegaDetector/SpeciesNet Categories in Images}
\usage{
megadet_get_counts(
  json,
  type = "detection",
  map_names = TRUE,
  validate_json = getOption("camRa.validate_json", default = TRUE)
)
}
\arguments{
\item{json}{character vector or nested list object from \code{\link[jsonlite:read_json]{jsonlite::read_json()}}.
The JSON file or loaded JSON data to filter on.}

\item{type}{character. "detection" if counts should be done from MegaDetector's gross
categories or "classification" for counts from SpeciesNet's species-level classifications.}

\item{map_names}{boolean. If MegaDetector and SpeciesNet names should be mapped from
numeric values to the values provided in the \code{detection_categories} and \code{classification_categories}
tags. These are typically gross (human, animal, vehicle) and taxonomic classifications respectively.}

\item{validate_json}{boolean. If JSON data formatted as nested lists should be validated.
This can prevent unexpected errors if the parameter is a list, but not JSON but may increase runtime.}
}
\value{
a dataframe with a category and count column.
}
\description{
Counts instances of each class for MegaDetector/SpeciesNet across images in JSON data.
}
