% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaDetector.R
\name{megadet_get_info}
\alias{megadet_get_info}
\title{Get Values from MegaDetector/SpeciesNet JSON Files}
\usage{
megadet_get_info(
  json,
  key = "info",
  file = NA,
  validate_json = getOption("camRa.validate_json", default = TRUE)
)
}
\arguments{
\item{json}{character vector or nested list object from \code{\link[jsonlite:read_json]{jsonlite::read_json()}}.
The JSON file or loaded JSON data to filter on.}

\item{key}{character, character vector, list, or \code{NULL}. Key(s) and/or indices to filter on in the JSON data.
For first level keys, provide a character of the key name. For higher levels, provide a vector
of keys in the order they are nested, such as \code{c("info", "megadetector_version")} to get the
MegaDetector version. A list with keys and indices can also be used, such as \code{list("images", 1)}
for detection information on the first image in the file. If \code{NULL} is passed as a key, the entire JSON
file will be returned.}

\item{file}{character. File to write JSON data to. Use \code{NA} to skip writing to a file.}

\item{validate_json}{boolean. If JSON data formatted as nested lists should be validated.
This can prevent unexpected errors if the parameter is a list, but not JSON but may increase runtime.}
}
\value{
dependent on keys used.
}
\description{
Grabs values from MegaDetector/SpeciesNet JSON files based on provided keys.
}
\examples{
\dontrun{
#Get data for first images in file
megadet_get_info(
  json = "classifications.json",
  key = list("images", 1)
)

#Get name of detector used
megadet_get_info(
  json = "classifications.json",
  key = c("info", "detector")
)

#Load entire JSON from file
megadet_get_info(
  json = "classifications.json",
  key = NULL
)
}

}
