% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{nacti_subset}
\alias{nacti_subset}
\title{Subset of North American Camera Trap Images files}
\format{
An object of class \code{data.frame} with 500 rows and 14 columns.
}
\usage{
nacti_subset
}
\description{
A subset of 500 sequential image files from the nacti dataset to be used for testing.
}
\keyword{datasets}
