% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaDetector.R
\name{specnet_reclassify}
\alias{specnet_reclassify}
\title{Update Classification Information in SpeciesNet JSON Data}
\usage{
specnet_reclassify(
  json,
  values_from,
  values_to,
  values_description = NA,
  file = NA,
  validate_json = getOption("camRa.validate_json", default = TRUE)
)
}
\arguments{
\item{json}{character vector or nested list object from \code{\link[jsonlite:read_json]{jsonlite::read_json()}}.
The JSON file or loaded JSON data to filter on.}

\item{values_from}{character vector. A vector of classification names currently in the JSON to
be updated. This must be unique values and include all names present in the file.}

\item{values_to}{character vector. A vector of classification names to reassign current classification
names as. Multiple copies of the same name can be present here to map values as many-to-one.}

\item{values_description}{character vector. A vector of classification descriptions to reassign current
classification descriptions as. This must have unique values to map onto \code{values_from}. Use \code{NA}
to skip updating descriptions and instead remove them from the file.}

\item{file}{character. File to write JSON data to. Use \code{NA} to skip writing to a file.}

\item{validate_json}{boolean. If JSON data formatted as nested lists should be validated.
This can prevent unexpected errors if the parameter is a list, but not JSON but may increase runtime.}
}
\value{
a nested list object representing JSON data.
}
\description{
Reclassifies classification categories in SpeciesNet JSON data with new species values,
including merging categories into one. This function is best used with a data frame mapping
old values to new ones. A list of current classification names can be pulled using
\code{\link[=megadet_get_info]{megadet_get_info()}} with \code{key = "classification_categories"} to build this table.
}
\examples{
\dontrun{
#CSV with columns for old values, new values, and new descriptions columns
value_map <- read.csv("mapped_values.csv")

specnet_reclassify(
  json = "classifications.json",
  values_from = value_map$old_value,
  values_to = value_map$new_value,
  values_description = value_map$new_description
)
}

}
