% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timelapse.R
\name{tl_ddb_extract}
\alias{tl_ddb_extract}
\title{Get Data Table from Timelapse Database}
\usage{
tl_ddb_extract(file)
}
\arguments{
\item{file}{character. File path to a timelapse database file (.ddb).}
}
\value{
a data frame.
}
\description{
Pulls the data table from a .ddb file used by the Timelapse program. Similar to the export
to CSV option in the GUI.
}
