% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ena24detection_subset}
\alias{ena24detection_subset}
\title{Random Subset of ena24detection files}
\format{
An object of class \code{data.frame} with 500 rows and 4 columns.
}
\usage{
ena24detection_subset
}
\description{
A random subset of 500 image files from the ena24detection dataset to be used for testing.
}
\keyword{datasets}
